requirements.txt
Flask==2.3.3
openai==0.27.0
python-dotenv==1.0.0
